+++ lib/swipe_reveal_card.dart
@@ -3,3 +3,5 @@ library;
 
 export 'src/swipe_action.dart';
 export 'src/swipe_reveal_card.dart';
+export 'src/swipe_reveal_controller.dart';
+export 'src/swipe_reveal_group.dart';