+++ test/step/i_swipe_the_card_left.dart
@@ -4,5 +4,6 @@ import 'package:flutter_test/flutter_test.dart';
 /// Usage: I swipe the card left
 Future<void> iSwipeTheCardLeft(WidgetTester tester) async {
   await tester.drag(find.byType(SingleChildScrollView), const Offset(-220, 0));
-  await tester.pumpAndSettle();
+  await tester.pump();
+  await tester.pump(const Duration(milliseconds: 300));
 }